pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Carlos Ortega Vazquez <ortegav_ce@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9"
hopsworks = "^3.4.2"
pandas = "^1.5.2"
jupyter = "^1.0.0"
tqdm = "^4.64.1"
scikit-learn = "^1.3.1"
xgboost = "^2.0.0"
optuna = "^3.4.0"
comet-ml = "^3.35.3"
python-dotenv = "^1.0.0"
streamlit = "^1.27.2"
tomlkit = "^0.12.4"
fire = "^0.5.0"
discordwebhook = "^1.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
